MW-199 Add donating and withdrawing money (virtual) (#42)
* Add donating and withdrawing money (virtual)

* Change AccountBalance's type from double to decimal

---------

+++ Entities/Data/User/UserDto.cs
@@ -24,7 +24,7 @@ namespace Entities.Data.User
 		///     Account balance
 		/// </summary>
 		/// <example>132</example>
-		public double? AccountBalance { get; set; }
+		public decimal? AccountBalance { get; set; }
 
 		/// <summary>
 		///     Subscriptions counter
